First pass at implementing v8::Message
v8::Message is expected to provide information about where an exception came from
which is used to print out information on unhandled exceptions like so:

E:\foo.js:2
@ <line of source code>
^
SyntaxError: Invalid or unexpected token

This change adds a JsrtExceptionMetadata class which knows the filename, line/column
where the error occurred, and the actual line of source which caused the error.

Currently I still need to add support for the filename in the case of compile errors
and I've noticed some differences in behavior around eval.

I also still need to add TTD support for this scenario.